+++ src/index.tsx
@@ -5,7 +5,7 @@ import './scss/index.scss';
 import './scss/elements.scss';
 import './scss/colors.scss';
 import { Provider } from 'react-redux';
-import { storeConfig } from './store';
+import { storeConfig } from './shared/api/store';
 
 const root = ReactDOM.createRoot(
   document.getElementById('root') as HTMLElement
