+++ src/pages/Profile/Profile.tsx
@@ -7,14 +7,14 @@ import empty from "../../images/empty-profile.png";
 import radiumLogo from "../../images/радиум лого.svg";
 import { IProfile } from "../../interfaces/user.interface";
 import styles from './Profile.module.scss';
-import { useAppDispatch, useAppSelector } from "../../hooks/redux";
+import { useAppDispatch, useAppSelector } from "@/shared/api/store";
 import Background from "../../ui/Background/Background";
 import TopPanel from "../../ui/TopPanel/TopPanel";
 import * as Icons from "../../shared/ui/icons";
 import List from "../../ui/List/List";
 import line from '../../images/Rectangle 11.svg'
 import { useDispatch } from "react-redux";
-import { fetchUser, updateProfileUser } from "../../store/actionCreators/actionCreatorsAuth";
+import { fetchUser, updateProfileUser } from "@/shared/api/actionCreators/actionCreatorsAuth";
 
 const Profile: FC = () => {
     const userName = 'андрей'
