+++ src/pages/RegisterPage/RegisterPage.tsx
@@ -3,8 +3,8 @@ import { SubmitHandler, useForm } from 'react-hook-form';
 import { Link, useNavigate } from 'react-router-dom';
 import Button from '../../ui/Button/Button';
 // import Error from "../../components/Error/Error";
-import { useAppDispatch } from '../../hooks/redux';
-import { registration } from '../../store/actionCreators/actionCreatorsAuth';
+import { useAppDispatch } from '@/shared/api/store';
+import { registration } from '@/shared/api/actionCreators/actionCreatorsAuth';
 import { IUser } from '../../interfaces/user.interface';
 import styles from './RegisterPage.module.scss';
 import { emailValidator } from '../../constData';
