+++ src/shared/api/store.ts
@@ -4,6 +4,7 @@ import courseReducer from './slices/courseSlice';
 import profileReducer from './slices/profileSlice';
 import moduleReducer from './slices/moduleSlice';
 import pageReducer from './slices/pageSlice';
+import {TypedUseSelectorHook, useDispatch, useSelector} from "react-redux";
 
 const rootReducer = combineReducers({
     auth: authReducer,
@@ -20,3 +21,6 @@ export const storeConfig = configureStore({
 export type RootState = ReturnType<typeof rootReducer>;
 export type AppStore = typeof storeConfig;
 export type AppDispatch = typeof storeConfig.dispatch;
+
+export const useAppSelector: TypedUseSelectorHook<RootState> = useSelector;
+export const useAppDispatch: () => AppDispatch = useDispatch;
